refactor(ramps): colocate buy pages with route entrypoints
Move build-quote and token-selection out of screens/ so route barrels
export real page components, matching shield/plan-style page layout.
